+++ src/flextype/core/Middlewares/CsrfMiddleware.php
@@ -9,8 +9,9 @@ declare(strict_types=1);
 
 namespace Flextype\Middlewares;
 
-use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
+use Psr\Http\Server\RequestHandlerInterface as RequestHandler;
+use Slim\Psr7\Response;
 
 class CsrfMiddleware
 {
@@ -21,7 +22,7 @@ class CsrfMiddleware
      * @param Response $response PSR7 response
      * @param callable $next     Next middleware
      */
-    public function __invoke(Request $request, Response $response, callable $next) : Response
+    /*public function __invoke(Request $request, Response $response, callable $next) : Response
     {
         $data = $request->getParsedBody();
 
@@ -33,6 +34,33 @@ class CsrfMiddleware
             }
         } else {
             $response = $next($request, $response);
+        }
+
+        return $response;
+    }*/
+
+
+    /**
+     * Invoke
+     *
+     * @param  ServerRequest  $request PSR-7 request
+     * @param  RequestHandler $handler PSR-15 request handler
+     *
+     * @return Response
+     */
+    public function __invoke(Request $request, RequestHandler $handler): Response
+    {
+        $response = $handler->handle($request);
+        $data = $request->getParsedBody();
+    
+        if (isset($data[csrf()->getTokenName()])) {
+            if (csrf()->isValid($data[csrf()->getTokenName()])) {
+                return $response;
+            } else {
+                $response = new Response();
+                $response->getBody()->write('This looks like a cross-site request forgery!');
+                return $response;
+            }
         }
 
         return $response;